Create A New endGame()  Function to Isolate a Game Ending
There are three prominent bugs: 1. The canGameContinue function
is unable to check for scenarios in which three or four tiles
are needed to equal the roll. 2. The canGameContinue function
sometimes does not appear to work as expected and the User will
be asked to enter a number when the remaining tiles will not
sum to equal the dice roll. 3. As this code uses recursion to
play each round, the call stack builds and needs to unwind
after the User no longer wants to play.